+++ _import/tycho-execution/foundry/scripts/deploy-executors.js
@@ -91,6 +91,13 @@ const executors_to_deploy = {
             exchange: "HashflowExecutor",
             args: ["0x55084eE0fEf03f14a305cd24286359A35D735151", "0x000000000022D473030F116dDEE9F6B43aC78BA3"]
         },
+        // Args: liquidity, permit2
+        {
+            exchange: "FluidV1Executor", args: [
+                "0x52Aa899454998Be5b000Ad077a46Bbe360F4e497",
+                "0x000000000022D473030F116dDEE9F6B43aC78BA3"
+            ]
+        },
     ],
     "base": [
         // Args: Factory, Pool Init Code Hash, Permit2, Fee BPS
@@ -110,6 +117,15 @@ const executors_to_deploy = {
                 "0x000000000022D473030F116dDEE9F6B43aC78BA3",
                 25
             ]
+        },
+        // PANCAKESWAP V3 - Args: Deployer, Pool Init Code Hash, Permit2
+        {
+            exchange: "UniswapV3Executor",
+            args: [
+                "0x41ff9AA7e16B8B1a8a8dc4f0eFacd93D02d071c9",
+                "0x6ce8eb472fa82df5469c6ab6d485f17c3ad13c8cd7af59b3d4a8026c5ce0f7e2",
+                "0x000000000022D473030F116dDEE9F6B43aC78BA3",
+            ],
         },
         // USV3 - Args: Factory, Pool Init Code Hash, Permit2
         {