+++ _import/tycho-execution/foundry/test/TychoRouterSequentialSwap.t.sol
@@ -516,6 +516,9 @@ contract TychoRouterSequentialSwapTest is TychoRouterTestSetup {
 }
 
 contract TychoRouterSequentialSwapTestForBebop is TychoRouterTestSetup {
+    // old BOB address with drainer contract here to avoid changing hardcoded test fixtures
+    address BOB_OLD = makeAddr("bob"); //bob=someone!=us
+
     function getForkBlock() public pure override returns (uint256) {
         return 23139046;
     }
@@ -530,17 +533,17 @@ contract TychoRouterSequentialSwapTestForBebop is TychoRouterTestSetup {
 
         uint256 amountIn = 1 ether;
         uint256 expectedAmountOut = 1672307;
-        deal(WETH_ADDR, BOB, amountIn);
-        uint256 balanceBefore = IERC20(WBTC_ADDR).balanceOf(BOB);
+        deal(WETH_ADDR, BOB_OLD, amountIn);
+        uint256 balanceBefore = IERC20(WBTC_ADDR).balanceOf(BOB_OLD);
 
-        vm.startPrank(BOB);
+        vm.startPrank(BOB_OLD);
         IERC20(WETH_ADDR).approve(tychoRouterAddr, type(uint256).max);
         bytes memory callData = loadCallDataFromFile("test_uniswap_v3_bebop");
         (bool success,) = tychoRouterAddr.call(callData);
 
         vm.stopPrank();
 
-        uint256 balanceAfter = IERC20(WBTC_ADDR).balanceOf(BOB);
+        uint256 balanceAfter = IERC20(WBTC_ADDR).balanceOf(BOB_OLD);
 
         assertTrue(success, "Call Failed");
         assertEq(balanceAfter - balanceBefore, expectedAmountOut);