+++ _import/tycho-execution/foundry/test/TychoRouterTestSetup.sol
@@ -15,6 +15,7 @@ import {
     IUniswapV3Pool
 } from "../src/executors/UniswapV3Executor.sol";
 import {UniswapV4Executor} from "../src/executors/UniswapV4Executor.sol";
+import {FluidV1Executor} from "../src/executors/FluidV1Executor.sol";
 import {SlipstreamsExecutor} from "../src/executors/SlipstreamsExecutor.sol";
 
 // Test utilities and mocks
@@ -78,6 +79,7 @@ contract TychoRouterTestSetup is Constants, Permit2TestHelper, TestUtils {
     BalancerV3Executor public balancerV3Executor;
     BebopExecutor public bebopExecutor;
     HashflowExecutor public hashflowExecutor;
+    FluidV1Executor public fluidV1Executor;
     SlipstreamsExecutor public slipstreamsExecutor;
 
     function getChain() public view virtual returns (string memory) {
@@ -146,10 +148,12 @@ contract TychoRouterTestSetup is Constants, Permit2TestHelper, TestUtils {
         bebopExecutor = new BebopExecutor(BEBOP_SETTLEMENT, PERMIT2_ADDRESS);
         hashflowExecutor =
             new HashflowExecutor(HASHFLOW_ROUTER, PERMIT2_ADDRESS);
+        fluidV1Executor =
+            new FluidV1Executor(FLUIDV1_LIQUIDITY, PERMIT2_ADDRESS);
         slipstreamsExecutor =
             new SlipstreamsExecutor(SLIPSTREAMS_FACTORY_BASE, PERMIT2_ADDRESS);
 
-        address[] memory executors = new address[](12);
+        address[] memory executors = new address[](13);
         executors[0] = address(usv2Executor);
         executors[1] = address(usv3Executor);
         executors[2] = address(pancakev3Executor);
@@ -161,7 +165,8 @@ contract TychoRouterTestSetup is Constants, Permit2TestHelper, TestUtils {
         executors[8] = address(balancerV3Executor);
         executors[9] = address(bebopExecutor);
         executors[10] = address(hashflowExecutor);
-        executors[11] = address(slipstreamsExecutor);
+        executors[11] = address(fluidV1Executor);
+        executors[12] = address(slipstreamsExecutor);
 
         return executors;
     }